# openapi.yaml
openapi: 3.0.1
info:
  title: Bio SDK Service
  version: 1.0.0
...
paths:
...
  /api/gallery/update-images-dir:
    post:
      summary: to update the current images dir
      operationId: updateImagesDir
      description: does a validation check on the provided path and if valid sets the images directory property of the SDKCore
      tags:
        - images-dir
        - config-help
      requestBody:
        description: Contains the new path for the images dir
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/models.updateImagesDirReqObj'
      responses:
        '200':
          description: Successful response
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                type: string
      deprecated: false
      security: []
  /api/gallery/update-cbor-dir:
    post:
      summary: to update the current cbor dir
      operationId: updateCborDir
      description: does a validation check on the provided path and if valid sets the cbor directory property of the SDKCore
      tags:
        - cbor-dir
        - config-help
      requestBody:
        description: Contains the new path for the cbor dir
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/models.updateCborDirReqObj'
      responses:
        '200':
          description: Successful response
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                type: string
      deprecated: false
      security: []
...
  /api/gallery/upload-cbor-zip:
    post:
      summary: used to upload a zip file containing cbor format templates
      operationId: uploadCborZip
      description: using multipart format zip file can be uploaded, once uploaded it will be unzipped and the zip will be deleted. Successful response contains success message along with file path to unzipped folder. 
      tags:
      - cbor-dir
      - config-help
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Successful response
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                type: string
      deprecated: false
      security: []
...
components:
  securitySchemes:
    default:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://test.com
          scopes: 
            read:books: Grants read access
  schemas:
    models.updateImagesDirReqObj:
      type: object
      properties:
        pathstring:
          type: string
          example: /app/imagesdir/
    models.updateCborDirReqObj:
      type: object
      properties:
        pathstring:
          type: string
          example: /app/cbordir/
...